pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "satnogs-mcp"
version = "0.1.0"
description = "MCP server for the SatNOGS open satellite observation network"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0",
    "requests>=2.31.0",
    "python-dotenv>=1.1.0",
    "redis>=5.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
satnogs-mcp = "satnogs_mcp.__main__:main"

[project.urls]
Homepage = "https://github.com/atlas/satnogs-mcp"
Repository = "https://github.com/atlas/satnogs-mcp"

[tool.hatch.build.targets.wheel]
packages = ["src/satnogs_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
